Moved entities and repositories to lib folder
