+++ apps/lib/src/main/java/com/cloudsherpa/lib/repositories/NormalizedMetricsRepository.java
@@ -1,7 +1,7 @@
-// Refer to EnvironmentReferenceRepository.java for most documentation
-package com.cloudsherpa.ingestion.normalization.persistence.repository;
+package com.cloudsherpa.lib.repositories;
 
-import com.cloudsherpa.ingestion.normalization.persistence.entity.NormalizedMetrics;
+import com.cloudsherpa.lib.entities.NormalizedMetrics;
+
 import java.time.OffsetDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -11,4 +11,7 @@ public interface NormalizedMetricsRepository extends JpaRepository<NormalizedMet
   // Spring translates this method name into:
   // SELECT * FROM normalized_metrics WHERE recorded_at BETWEEN ? AND ?
   List<NormalizedMetrics> findByRecordedAtBetween(OffsetDateTime startTime, OffsetDateTime endTime);
+
+  List<NormalizedMetrics> findByPeriodStartBetween(
+      OffsetDateTime startTime, OffsetDateTime endTime);
 }
