+++ apps/service/pom.xml
@@ -95,6 +95,11 @@
 			<artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>
 			<version>2.8.17</version>
 		</dependency>
+		<dependency>
+			<groupId>com.cloudsherpa</groupId>
+			<artifactId>lib</artifactId>
+			<version>0.0.1-SNAPSHOT</version>
+		</dependency>
 	</dependencies>
 
 	<build>
