--- apps/service/src/main/java/com/cloudsherpa/service/analytics/repository/NormalizedMetricsRepository.java
@@ -1,15 +1,0 @@
-// Refer to EnvironmentReferenceRepository.java for most documentation
-package com.cloudsherpa.service.analytics.repository;
-
-import com.cloudsherpa.service.analytics.entities.NormalizedMetrics;
-import java.time.OffsetDateTime;
-import java.util.List;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface NormalizedMetricsRepository extends JpaRepository<NormalizedMetrics, UUID> {
-  // Spring translates this method name into:
-  // SELECT * FROM normalized_metrics WHERE recorded_at BETWEEN ? AND ?
-  List<NormalizedMetrics> findByPeriodStartBetween(
-      OffsetDateTime startTime, OffsetDateTime endTime);
-}
