+++ apps/service/src/main/java/com/cloudsherpa/service/auth/service/AuthService.java
@@ -1,10 +1,10 @@
 package com.cloudsherpa.service.auth.service;
 
+import com.cloudsherpa.lib.entities.User;
+import com.cloudsherpa.lib.repositories.UserRepository;
 import com.cloudsherpa.service.auth.dto.AuthUserResponse;
 import com.cloudsherpa.service.auth.dto.LoginRequest;
 import com.cloudsherpa.service.auth.dto.RegisterRequest;
-import com.cloudsherpa.service.auth.model.User;
-import com.cloudsherpa.service.auth.repository.UserRepository;
 import java.util.UUID;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
